+++ x/ibc-hooks/client/cli/query.go
@@ -5,7 +5,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/version"
-	ibchooks "github.com/osmosis-labs/osmosis/x/ibc-hooks"
+	"github.com/osmosis-labs/osmosis/x/ibc-hooks/keeper"
 	"github.com/spf13/cobra"
 	"strings"
 
@@ -41,7 +41,7 @@ $ %s query ibc-hooks wasm-hooks-sender channel-42 juno12smx2wdlyttvyzvzg54y2vnqw
 			originalSender := args[1]
 			// ToDo: Make this flexible as an arg
 			prefix := sdk.GetConfig().GetBech32AccountAddrPrefix()
-			senderBech32, err := ibchooks.DeriveIntermediateSender(channelID, originalSender, prefix)
+			senderBech32, err := keeper.DeriveIntermediateSender(channelID, originalSender, prefix)
 			if err != nil {
 				return err
 			}
